#4625 - Azure AI Open AI recommender
- Pull common code out into a support module
