+++ inception/inception-bom/pom.xml
@@ -289,6 +289,11 @@
       </dependency>
       <dependency>
         <groupId>de.tudarmstadt.ukp.inception.app</groupId>
+        <artifactId>inception-imls-support-llm</artifactId>
+        <version>32.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>de.tudarmstadt.ukp.inception.app</groupId>
         <artifactId>inception-imls-weblicht</artifactId>
         <version>32.0-SNAPSHOT</version>
       </dependency>
