+++ inception/inception-imls-azureai-openai/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/azureaiopenai/Preset.java
@@ -22,8 +22,8 @@ import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import de.tudarmstadt.ukp.inception.recommendation.imls.azureaiopenai.client.GenerateResponseFormat;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.ExtractionMode;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt.PromptingMode;
+import de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.prompt.PromptingMode;
+import de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.response.ExtractionMode;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Preset
