+++ inception/inception-imls-azureai-openai/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/azureaiopenai/client/ChatCompletionRequest.java
@@ -27,7 +27,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.client.Option;
+import de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.option.Option;
 
 public class ChatCompletionRequest
 {
