+++ inception/inception-imls-ollama/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/ollama/OllamaRecommender.java
@@ -17,22 +17,18 @@
  */
 package de.tudarmstadt.ukp.inception.recommendation.imls.ollama;
 
-import static de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt.PromptContextGenerator.VAR_EXAMPLES;
+import static de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.prompt.PromptContextGenerator.VAR_EXAMPLES;
+import static de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.prompt.PromptContextGenerator.getPromptContextGenerator;
+import static de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.response.ResponseExtractor.getResponseExtractor;
 
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
-import java.nio.charset.Charset;
+import java.util.Map;
 
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.uima.cas.CAS;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import com.hubspot.jinjava.Jinjava;
-import com.hubspot.jinjava.JinjavaConfig;
-import com.hubspot.jinjava.interpret.JinjavaInterpreter;
-import com.hubspot.jinjava.loader.ResourceLocator;
-import com.hubspot.jinjava.loader.ResourceNotFoundException;
 
 import de.tudarmstadt.ukp.inception.recommendation.api.model.Recommender;
 import de.tudarmstadt.ukp.inception.recommendation.api.recommender.NonTrainableRecommenderEngineImplBase;
@@ -40,14 +36,7 @@ import de.tudarmstadt.ukp.inception.recommendation.api.recommender.PredictionCon
 import de.tudarmstadt.ukp.inception.recommendation.api.recommender.RecommendationException;
 import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.client.OllamaClient;
 import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.client.OllamaGenerateRequest;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt.PerAnnotationContextGenerator;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt.PerDocumentContextGenerator;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt.PerSentenceContextGenerator;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt.PromptContext;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt.PromptContextGenerator;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.response.MentionsFromJsonExtractor;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.response.ResponseAsLabelExtractor;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.response.ResponseExtractor;
+import de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.prompt.JinjaPromptRenderer;
 import de.tudarmstadt.ukp.inception.rendering.model.Range;
 import de.tudarmstadt.ukp.inception.support.logging.LogMessage;
 
@@ -61,7 +50,8 @@ public class OllamaRecommender
     private final OllamaRecommenderTraits traits;
 
     private final OllamaClient client;
-    private final Jinjava jinjava;
+
+    private final JinjaPromptRenderer promptRenderer;
 
     public OllamaRecommender(Recommender aRecommender, OllamaRecommenderTraits aTraits,
             OllamaClient aClient)
@@ -69,60 +59,43 @@ public class OllamaRecommender
         super(aRecommender);
 
         traits = aTraits;
-
-        var config = new JinjavaConfig();
-        jinjava = new Jinjava(config);
-        jinjava.setResourceLocator(new ResourceLocator()
-        {
-            @Override
-            public String getString(String aFullName, Charset aEncoding,
-                    JinjavaInterpreter aInterpreter)
-                throws IOException
-            {
-                throw new ResourceNotFoundException("Couldn't find resource: " + aFullName);
-            }
-        });
-
         client = aClient;
+        promptRenderer = new JinjaPromptRenderer();
     }
 
     @Override
     public Range predict(PredictionContext aContext, CAS aCas, int aBegin, int aEnd)
         throws RecommendationException
     {
-        var responseExtractor = getResponseExtractor();
+        var responseExtractor = getResponseExtractor(traits.getExtractionMode());
         var examples = responseExtractor.generate(this, aCas, MAX_FEW_SHOT_EXAMPLES);
+        var globalBindings = Map.of(VAR_EXAMPLES, examples);
 
-        getPromptContextGenerator().generate(this, aCas, aBegin, aEnd).forEach(promptContext -> {
-            try {
-                var bindings = promptContext.getBindings();
-
-                bindings.put(VAR_EXAMPLES, examples);
-
-                var response = query(promptContext);
-
-                responseExtractor.extract(this, aCas, promptContext, response);
-            }
-            catch (IOException e) {
-                aContext.log(LogMessage.warn(getRecommender().getName(),
-                        "Ollama [%s] failed to respond: %s", traits.getModel(),
-                        ExceptionUtils.getRootCauseMessage(e)));
-                LOG.error("Ollama [{}] failed to respond: {}", traits.getModel(),
-                        ExceptionUtils.getRootCauseMessage(e));
-            }
-        });
+        getPromptContextGenerator(traits.getPromptingMode())
+                .generate(this, aCas, aBegin, aEnd, globalBindings).forEach(promptContext -> {
+                    try {
+                        var prompt = promptRenderer.render(traits.getPrompt(), promptContext);
+                        var response = query(prompt);
+                        responseExtractor.extract(this, aCas, promptContext, response);
+                    }
+                    catch (IOException e) {
+                        aContext.log(LogMessage.warn(getRecommender().getName(),
+                                "Ollama [%s] failed to respond: %s", traits.getModel(),
+                                ExceptionUtils.getRootCauseMessage(e)));
+                        LOG.error("Ollama [{}] failed to respond: {}", traits.getModel(),
+                                ExceptionUtils.getRootCauseMessage(e));
+                    }
+                });
 
         return new Range(aBegin, aEnd);
     }
 
-    private String query(PromptContext aContext) throws IOException
+    private String query(String aPrompt) throws IOException
     {
-        var prompt = jinjava.render(traits.getPrompt(), aContext.getBindings());
-
-        LOG.trace("Querying ollama [{}]: [{}]", traits.getModel(), prompt);
+        LOG.trace("Querying ollama [{}]: [{}]", traits.getModel(), aPrompt);
         var request = OllamaGenerateRequest.builder() //
                 .withModel(traits.getModel()) //
-                .withPrompt(prompt) //
+                .withPrompt(aPrompt) //
                 .withFormat(traits.getFormat()) //
                 .withRaw(traits.isRaw()) //
                 .withStream(false) //
@@ -132,33 +105,5 @@ public class OllamaRecommender
         var response = client.generate(traits.getUrl(), request).trim();
         LOG.trace("Ollama [{}] responds: [{}]", traits.getModel(), response);
         return response;
-    }
-
-    private PromptContextGenerator getPromptContextGenerator()
-    {
-        switch (traits.getPromptingMode()) {
-        case PER_ANNOTATION:
-            return new PerAnnotationContextGenerator();
-        case PER_SENTENCE:
-            return new PerSentenceContextGenerator();
-        case PER_DOCUMENT:
-            return new PerDocumentContextGenerator();
-        default:
-            throw new IllegalArgumentException(
-                    "Unsupported mode [" + traits.getPromptingMode() + "]");
-        }
-    }
-
-    private ResponseExtractor getResponseExtractor()
-    {
-        switch (traits.getExtractionMode()) {
-        case RESPONSE_AS_LABEL:
-            return new ResponseAsLabelExtractor();
-        case MENTIONS_FROM_JSON:
-            return new MentionsFromJsonExtractor();
-        default:
-            throw new IllegalArgumentException(
-                    "Unsupported extraction mode [" + traits.getExtractionMode() + "]");
-        }
     }
 }
