+++ inception/inception-imls-ollama/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/ollama/OllamaRecommenderTraits.java
@@ -28,7 +28,8 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.client.OllamaGenerateResponseFormat;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt.PromptingMode;
+import de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.prompt.PromptingMode;
+import de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.response.ExtractionMode;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class OllamaRecommenderTraits
