+++ inception/inception-imls-ollama/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/ollama/Preset.java
@@ -22,7 +22,8 @@ import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.client.OllamaGenerateResponseFormat;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt.PromptingMode;
+import de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.prompt.PromptingMode;
+import de.tudarmstadt.ukp.inception.recommendation.imls.support.llm.response.ExtractionMode;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Preset
