--- inception/inception-imls-ollama/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/ollama/prompt/PerAnnotationContextGenerator.java
@@ -1,50 +1,0 @@
-/*
- * Licensed to the Technische Universität Darmstadt under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The Technische Universität Darmstadt 
- * licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.
- *  
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt;
-
-import static de.tudarmstadt.ukp.inception.support.uima.WebAnnoCasUtil.selectOverlapping;
-
-import java.util.stream.Stream;
-
-import org.apache.uima.cas.CAS;
-
-import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence;
-import de.tudarmstadt.ukp.inception.recommendation.api.recommender.RecommendationEngine;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.jinjava.CasWrapper;
-
-public class PerAnnotationContextGenerator
-    implements PromptContextGenerator
-{
-    @Override
-    public Stream<PromptContext> generate(RecommendationEngine aEngine, CAS aCas, int aBegin,
-            int aEnd)
-    {
-        var predictedType = aEngine.getPredictedType(aCas);
-        var candidates = selectOverlapping(aCas, predictedType, aBegin, aEnd);
-        return candidates.stream().map(candidate -> {
-            var sentence = aCas.select(Sentence.class).covering(candidate) //
-                    .map(Sentence::getCoveredText) //
-                    .findFirst().orElse("");
-            var context = new PromptContext(candidate);
-            context.set(VAR_CAS, new CasWrapper(aCas));
-            context.set(VAR_TEXT, candidate.getCoveredText());
-            context.set(VAR_SENTENCE, sentence);
-            return context;
-        });
-    }
-}
