--- inception/inception-imls-ollama/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/ollama/prompt/PerSentenceContextGenerator.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Technische Universität Darmstadt under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The Technische Universität Darmstadt 
- * licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.
- *  
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.tudarmstadt.ukp.inception.recommendation.imls.ollama.prompt;
-
-import static de.tudarmstadt.ukp.inception.support.uima.WebAnnoCasUtil.selectOverlapping;
-
-import java.util.stream.Stream;
-
-import org.apache.uima.cas.CAS;
-import org.apache.uima.fit.util.CasUtil;
-
-import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence;
-import de.tudarmstadt.ukp.inception.recommendation.api.recommender.RecommendationEngine;
-import de.tudarmstadt.ukp.inception.recommendation.imls.ollama.jinjava.CasWrapper;
-
-public class PerSentenceContextGenerator
-    implements PromptContextGenerator
-{
-    @Override
-    public Stream<PromptContext> generate(RecommendationEngine aEngine, CAS aCas, int aBegin,
-            int aEnd)
-    {
-        var sentenceType = CasUtil.getAnnotationType(aCas, Sentence.class);
-
-        var candidates = selectOverlapping(aCas, sentenceType, aBegin, aEnd);
-
-        return candidates.stream().map(candidate -> {
-            var context = new PromptContext(candidate);
-            context.set(VAR_CAS, new CasWrapper(aCas));
-            context.set(VAR_TEXT, candidate.getCoveredText());
-            return context;
-        });
-    }
-}
