--- inception/inception-imls-ollama/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/ollama/response/MentionsSample.java
@@ -1,57 +1,0 @@
-/*
- * Licensed to the Technische Universität Darmstadt under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The Technische Universität Darmstadt 
- * licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.
- *  
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.tudarmstadt.ukp.inception.recommendation.imls.ollama.response;
-
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
-
-public class MentionsSample
-{
-    private final String text;
-    private final Map<String, String> labelledMentions = new LinkedHashMap<>();
-
-    public MentionsSample(String aText)
-    {
-        text = aText;
-    }
-
-    public void addMention(String aMention, String aLabel)
-    {
-        labelledMentions.put(aMention, aLabel);
-    }
-
-    public String getText()
-    {
-        return text;
-    }
-
-    public Map<String, String> getLabelledMentions()
-    {
-        return labelledMentions;
-    }
-
-    @Override
-    public String toString()
-    {
-        return new ToStringBuilder(this, ToStringStyle.SIMPLE_STYLE).append("text", text)
-                .append("labelledMentions", labelledMentions).toString();
-    }
-}
