+++ inception/pom.xml
@@ -257,7 +257,7 @@
     <module>inception-layer-docmetadata</module>
     <module>inception-feature-lookup</module>
     <module>inception-image</module>
-    <!-- Recommendation system -->
+    <!-- Recommender system -->
     <module>inception-recommendation-api</module>
     <module>inception-recommendation</module>
     <module>inception-active-learning</module>
@@ -269,6 +269,7 @@
     <module>inception-imls-external</module>
     <module>inception-imls-lapps</module>
     <module>inception-imls-ollama</module>
+    <module>inception-imls-support-llm</module>
     <module>inception-imls-weblicht</module>
     <!-- Documentation -->
     <module>inception-doc</module>
